Deprecated tmaptools functions

The following functions are not used anymore or deprecated as of tmaptools version 3.0. These functions are based on the sp package and are not supported anymore. They have been migrated to https://github.com/mtennekes/oldtmaptools

Usage

get_proj4(x, output = c("crs", "character", "epsg", "CRS"))

set_projection(
  shp,
  projection = NA,
  current.projection = NA,
  overwrite.current.projection = FALSE
)

get_projection(
  shp,
  guess.longlat = FALSE,
  output = c("character", "crs", "epsg", "CRS")
)

is_projected(x)

Arguments

x

see documentation in tmaptools 2.x for details

output

see documentation in tmaptools 2.x for details

shp

see documentation in tmaptools 2.x for details

projection

see documentation in tmaptools 2.x for details

current.projection

see documentation in tmaptools 2.x for details

overwrite.current.projection

see documentation in tmaptools 2.x for details

guess.longlat

see documentation in tmaptools 2.x for details

Deprecated as of version 2.0

append_data, aggregate_map, double_line, points_to_raster, poly_to_raster, sample_dots, sbind, smooth_map, smooth_raster_cover, read_shape, write_shape. These functions are based on the sp package and are not supported anymore. They have been migrated to https://github.com/mtennekes/oldtmaptools

osm_poly, osm_line, osm_point. Please use the package osmdata

Deprecated as of version 3.0

get_proj4: for projections, please use the function st_crs

set_projection for setting map projections, use st_crs or st_transform

get_projection for getting map projections, use st_crs

is_projected for checking if projections are long lat coordinates, use st_is_longlat