Explore color palettes

palette_explorer() starts an interactive tool shows all Color Brewer and viridis palettes, where the number of colors can be adjusted as well as the constrast range. Categorical (qualitative) palettes can be stretched when the number of colors exceeds the number of palette colors. Output code needed to get the desired color values is generated. Finally, all colors can be tested for color blindness. The data.frame tmap.pal.info is similar to brewer.pal.info, but extended with the color palettes from viridis.

Usage

palette_explorer()

tmap.pal.info

Format

An object of class data.frame with 40 rows and 4 columns.

References

https://www.color-blindness.com/types-of-color-blindness/

See also

Examples

if (FALSE) { # \dontrun{
if (require(shiny) && require(shinyjs)) {
    palette_explorer()
}
} # }