Thematic Map Tools

This package offers a set of handy tool functions for reading and processing spatial data. The aim of these functions is to supply the workflow to create thematic maps, e.g. read shape files, set map projections, append data, calculate areas and distances, and query OpenStreetMap. The visualization of thematic maps can be done with the tmap package.

Details

This page provides a brief overview of all package functions.

Tool functions (shape)

approx_areas	Approximate area sizes of polygons

approx_distances	Approximate distances

bb	Create, extract or modify a bounding box

bb_poly	Convert bounding box to a polygon

get_asp_ratio	Get the aspect ratio of a shape object

—————————	—————————————————————————————————

Tool functions (colors)

get_brewer_pal	Get and plot a (modified) Color Brewer palette

map_coloring	Find different colors for adjacent polygons

palette_explorer	Explore Color Brewer palettes

—————————	—————————————————————————————————

Spatial transformation functions

crop_shape	Crop shape objects

simplify_shape	Simplify a shape

—————————	—————————————————————————————————

Input and output functions

geocode_OSM	Get a location from an address description

read_GPX	Read a GPX file

read_osm	Read Open Street Map data

rev_geocode_OSM	Get an address description from a location

—————————	—————————————————————————————————

See also

Author

Maintainer: Martijn Tennekes mtennekes@gmail.com